Создан новый модуль filter.js
