+++ js/filter.js
@@ -1,0 +1,4 @@
+const formFilters = document.querySelector('.map__filters');
+const formFiltersContainer = Array.from(formFilters);
+
+export { formFilters, formFiltersContainer };
