+++ js/form.js
@@ -2,4 +2,3 @@ const form = document.querySelector('.ad-form');
 const formContainer = Array.from(form);
 
 export { form, formContainer };
-
