+++ js/main.js
@@ -1,12 +1,6 @@
 import { adverts } from './data.js';
-import { addCard, formFiltersContainer, formFilters } from './map.js';
-import { form, formContainer } from './form.js';
-import { disableForm, enableForm } from './state.js';
+import { addCard } from './map.js';
+import './page.js';
 
 addCard(adverts[6]);
 
-disableForm(formFilters, formFiltersContainer);
-disableForm(form, formContainer);
-
-enableForm(formFilters, formFiltersContainer);
-enableForm(form, formContainer);
