+++ js/map.js
@@ -1,8 +1,6 @@
 import { renderCard } from './card.js';
 
 const mapCanvas = document.querySelector('#map-canvas');
-const formFilters = document.querySelector('.map__filters');
-const formFiltersContainer = Array.from(formFilters);
 
 const addCard = (advert) => {
   const card = renderCard(advert);
@@ -10,4 +8,4 @@ const addCard = (advert) => {
   mapCanvas.appendChild(card);
 };
 
-export { addCard, formFilters, formFiltersContainer };
+export { addCard };
